#50 Move double functions to doubles header file.
