+++ graphics/coord.h
@@ -1,27 +1,13 @@
 #pragma once
 
 #include "transforms.h"
+#include "doubles.h"
 
 #include <vector>
 #include <list>
 #include <algorithm>
 
 using namespace std;
-
-#define UNUSED __attribute__ ((unused))
-
-// Absolute difference between a and b
-inline double abs_diff(double a, double b)
-{
-    return abs(a - b);
-}
-
-// Determine if a and b are equal, accepting up to epsilon as the difference.
-inline bool equals(double a, double b)
-{
-    constexpr double epsilon = 0.000001;
-    return abs_diff(a, b) < epsilon;
-}
 
 // 2D coordinates
 class Coord;
